Moved endpoints definition to a separate class
